moving the xml decoder to its own subpackage based on PR comments
